[SPARK-40064][SQL] Use V2 Filter in SupportsOverwrite
