+++ sql/catalyst/src/test/scala/org/apache/spark/sql/connector/catalog/InMemoryTable.scala
@@ -19,10 +19,14 @@ package org.apache.spark.sql.connector.catalog
 
 import java.util
 
+import org.scalatest.Assertions.assert
+
 import org.apache.spark.sql.connector.distributions.{Distribution, Distributions}
 import org.apache.spark.sql.connector.expressions.{SortOrder, Transform}
-import org.apache.spark.sql.sources.Filter
+import org.apache.spark.sql.connector.write.{LogicalWriteInfo, SupportsOverwrite, WriteBuilder, WriterCommitMessage}
+import org.apache.spark.sql.sources._
 import org.apache.spark.sql.types.StructType
+import org.apache.spark.sql.util.CaseInsensitiveStringMap
 
 /**
  * A simple in-memory table. Rows are stored as a buffered group produced by each output task.
@@ -40,12 +44,12 @@ class InMemoryTable(
     ordering, numPartitions, isDistributionStrictlyRequired) with SupportsDelete {
 
   override def canDeleteWhere(filters: Array[Filter]): Boolean = {
-    InMemoryBaseTable.supportsFilters(filters)
+    InMemoryTable.supportsFilters(filters)
   }
 
   override def deleteWhere(filters: Array[Filter]): Unit = dataMap.synchronized {
     import org.apache.spark.sql.connector.catalog.CatalogV2Implicits.MultipartIdentifierHelper
-    dataMap --= InMemoryBaseTable.filtersToKeys(dataMap.keys, partCols.map(_.toSeq.quoted), filters)
+    dataMap --= InMemoryTable.filtersToKeys(dataMap.keys, partCols.map(_.toSeq.quoted), filters)
   }
 
   override def withData(data: Array[BufferedRows]): InMemoryTable = {
@@ -63,5 +67,94 @@ class InMemoryTable(
       addPartitionKey(key)
     })
     this
+  }
+
+  override def newWriteBuilder(info: LogicalWriteInfo): WriteBuilder = {
+    InMemoryBaseTable.maybeSimulateFailedTableWrite(new CaseInsensitiveStringMap(properties))
+    InMemoryBaseTable.maybeSimulateFailedTableWrite(info.options)
+
+    new InMemoryWriterBuilderWithOverWrite()
+  }
+
+  private class InMemoryWriterBuilderWithOverWrite() extends InMemoryWriterBuilder
+    with SupportsOverwrite {
+
+    override def truncate(): WriteBuilder = {
+      assert(writer == Append)
+      writer = TruncateAndAppend
+      streamingWriter = StreamingTruncateAndAppend
+      this
+    }
+
+    override def overwrite(filters: Array[Filter]): WriteBuilder = {
+      assert(writer == Append)
+      writer = new Overwrite(filters)
+      streamingWriter = new StreamingNotSupportedOperation(
+        s"overwrite (${filters.mkString("filters(", ", ", ")")})")
+      this
+    }
+
+    override def canOverwrite(filters: Array[Filter]): Boolean = {
+      InMemoryTable.supportsFilters(filters)
+    }
+  }
+
+  private class Overwrite(filters: Array[Filter]) extends TestBatchWrite {
+    import org.apache.spark.sql.connector.catalog.CatalogV2Implicits.MultipartIdentifierHelper
+    override def commit(messages: Array[WriterCommitMessage]): Unit = dataMap.synchronized {
+      val deleteKeys = InMemoryTable.filtersToKeys(
+        dataMap.keys, partCols.map(_.toSeq.quoted), filters)
+      dataMap --= deleteKeys
+      withData(messages.map(_.asInstanceOf[BufferedRows]))
+    }
+  }
+}
+
+object InMemoryTable {
+
+  def filtersToKeys(
+      keys: Iterable[Seq[Any]],
+      partitionNames: Seq[String],
+      filters: Array[Filter]): Iterable[Seq[Any]] = {
+    keys.filter { partValues =>
+      filters.flatMap(splitAnd).forall {
+        case EqualTo(attr, value) =>
+          value == InMemoryBaseTable.extractValue(attr, partitionNames, partValues)
+        case EqualNullSafe(attr, value) =>
+          val attrVal = InMemoryBaseTable.extractValue(attr, partitionNames, partValues)
+          if (attrVal == null && value == null) {
+            true
+          } else if (attrVal == null || value == null) {
+            false
+          } else {
+            value == attrVal
+          }
+        case IsNull(attr) =>
+          null == InMemoryBaseTable.extractValue(attr, partitionNames, partValues)
+        case IsNotNull(attr) =>
+          null != InMemoryBaseTable.extractValue(attr, partitionNames, partValues)
+        case AlwaysTrue() => true
+        case f =>
+          throw new IllegalArgumentException(s"Unsupported filter type: $f")
+      }
+    }
+  }
+
+  def supportsFilters(filters: Array[Filter]): Boolean = {
+    filters.flatMap(splitAnd).forall {
+      case _: EqualTo => true
+      case _: EqualNullSafe => true
+      case _: IsNull => true
+      case _: IsNotNull => true
+      case _: AlwaysTrue => true
+      case _ => false
+    }
+  }
+
+  private def splitAnd(filter: Filter): Seq[Filter] = {
+    filter match {
+      case And(left, right) => splitAnd(left) ++ splitAnd(right)
+      case _ => filter :: Nil
+    }
   }
 }
